Fix Mac build break.
Review URL: http://codereview.chromium.org/56137

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12931 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/plugin/webplugin_proxy.cc
@@ -441,7 +441,7 @@ void WebPluginProxy::InitiateHTTPRangeRequest(const char* url,
                                               const char* range_info,
                                               intptr_t existing_stream,
                                               bool notify_needed,
-                                              int notify_data) {
+                                              intptr_t notify_data) {
 
   Send(new PluginHostMsg_InitiateHTTPRangeRequest(route_id_, url,
                                                   range_info, existing_stream,
